requirements/common.txt
alembic==1.10.4
async-generator==1.10
attrs==23.1.0
bleach==2.0.0
blinker==1.6.2
build==0.10.0
certifi==2023.5.7
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
dnspython==2.3.0
dominate==2.7.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
Faker==18.10.1
Flask==2.3.2
Flask-Bootstrap==3.3.7.1
Flask-Ext==0.1
Flask-HTTPAuth==3.2.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-Moment==1.0.5
Flask-PageDown==0.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
h11==0.14.0
html5lib==1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
Markdown==2.6.8
MarkupSafe==2.1.2
outcome==1.2.0
packaging==23.1
pip-tools==6.13.0
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.7.0
pyproject_hooks==1.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.6.5
python-editor==1.0.4
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.12
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.5.0
urllib3==2.0.3
visitor==0.1.3
webencodings==0.5.1
Werkzeug==2.3.3
wsproto==1.2.0
WTForms==3.0.1
